+++ tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/MetricName.java
@@ -16,13 +16,8 @@
 
 package com.palantir.tritium.metrics.registry;
 
-import com.google.common.collect.ImmutableMap;
-import com.google.errorprone.annotations.CanIgnoreReturnValue;
-import com.palantir.logsafe.Preconditions;
 import com.palantir.logsafe.Safe;
-import java.util.Map;
 import java.util.SortedMap;
-import javax.annotation.Nullable;
 
 public interface MetricName {
 
@@ -44,68 +39,5 @@ public interface MetricName {
         return new Builder();
     }
 
-    final class Builder {
-        @Nullable
-        private String safeName;
-
-        private TagMap tagMap = TagMap.EMPTY;
-
-        private Builder() {}
-
-        @CanIgnoreReturnValue
-        public Builder from(MetricName instance) {
-            Preconditions.checkNotNull(instance, "instance");
-            safeName(instance.safeName());
-            putAllSafeTags(instance.safeTags());
-            return this;
-        }
-
-        @CanIgnoreReturnValue
-        public Builder safeName(String value) {
-            this.safeName = Preconditions.checkNotNull(value, "safeName");
-            return this;
-        }
-
-        @CanIgnoreReturnValue
-        public Builder putSafeTags(String key, String value) {
-            Preconditions.checkNotNull(key, "safeTagName");
-            Preconditions.checkNotNull(value, "safeTagValue");
-            tagMap = tagMap.withEntry(key, value);
-            return this;
-        }
-
-        @CanIgnoreReturnValue
-        public Builder putSafeTags(Map.Entry<String, ? extends String> entry) {
-            Preconditions.checkNotNull(entry, "entry");
-            tagMap = tagMap.withEntry(entry.getKey(), entry.getValue());
-            return this;
-        }
-
-        @SuppressWarnings("unchecked")
-        @CanIgnoreReturnValue
-        public Builder safeTags(Map<String, ? extends String> entries) {
-            Preconditions.checkNotNull(entries, "entries");
-            tagMap = TagMap.of((Map<String, String>) entries);
-            return this;
-        }
-
-        @SuppressWarnings("unchecked")
-        @CanIgnoreReturnValue
-        public Builder putAllSafeTags(Map<String, ? extends String> entries) {
-            Preconditions.checkNotNull(entries, "entries");
-            if (!entries.isEmpty()) {
-                tagMap = tagMap.isEmpty()
-                        ? TagMap.of((Map<String, String>) entries)
-                        : TagMap.of(ImmutableMap.<String, String>builderWithExpectedSize(tagMap.size() + entries.size())
-                                .putAll(tagMap)
-                                .putAll(entries)
-                                .build());
-            }
-            return this;
-        }
-
-        public MetricName build() {
-            return new RealMetricName(Preconditions.checkNotNull(safeName, "safeName is required"), tagMap);
-        }
-    }
+    final class Builder extends ImmutableMetricName.Builder {}
 }
